Input handling fixed to work with GLFW

+++ OpenGL OGLDEV/headers/TimeStep.h
@@ -1,6 +1,6 @@
 #pragma once
 struct TimeStep {
 	TimeStep();
-	int timeInterval;
-	int lastTime;
+	double timeInterval;
+	double lastTime;
 };
